+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed map settings impossible to save or import if Navigraph Terminal Waypoints were enabled or disabled at some point
 - Fixed map settings impossible to save or import if Aircraft scale settings is set to anything with 2 number after dot (for example, 1.35)
 - Fixed some airways not showing as intended
+- Add ZAU Center and TRACONs to ATC duplication
 
 # [1.2.4-6]